+++ lib/generators/view_component/erb/erb_generator.rb
@@ -11,6 +11,7 @@ module ViewComponent
       source_root File.expand_path("templates", __dir__)
       class_option :sidecar, type: :boolean, default: false
       class_option :inline, type: :boolean, default: false
+      class_option :call, type: :boolean, default: false
       class_option :stimulus, type: :boolean, default: false
 
       def engine_name